Add react routes
