+++ frontend/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "framer-motion": "^12.23.5",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router-dom": "^7.7.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.30.1",
