+++ frontend/src/main.tsx
@@ -1,11 +1,14 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
+import { BrowserRouter } from 'react-router-dom';
 import './index.css'
 import App from './App.tsx'
 import Test_component from './Test.tsx'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
-    <App />
+     <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </StrictMode>,
 )
